+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 
 - More gui text has translation keys now, making them translatable by a resource pack
+- Fixed some cases of cruxtruder data changes not being saved
 
 ## [1.18.2-1.9.2.0] - 2023-03-13